feat: add url helpers and extract to url file
