+++ src/utils/request.ts
@@ -64,46 +64,3 @@ export function assertMethod(
     });
   }
 }
-
-export function getRequestHost(
-  event: H3Event,
-  opts: { xForwardedHost?: boolean } = {}
-) {
-  if (opts.xForwardedHost) {
-    const xForwardedHost = event.node.req.headers["x-forwarded-host"] as string;
-    if (xForwardedHost) {
-      return xForwardedHost;
-    }
-  }
-  return event.node.req.headers.host || "localhost";
-}
-
-export function getRequestProtocol(
-  event: H3Event,
-  opts: { xForwardedProto?: boolean } = {}
-) {
-  if (
-    opts.xForwardedProto !== false &&
-    event.node.req.headers["x-forwarded-proto"] === "https"
-  ) {
-    return "https";
-  }
-  return (event.node.req.connection as any).encrypted ? "https" : "http";
-}
-
-const DOUBLE_SLASH_RE = /[/\\]{2,}/g;
-
-export function getRequestPath(event: H3Event): string {
-  const path = (event.node.req.url || "/").replace(DOUBLE_SLASH_RE, "/");
-  return path;
-}
-
-export function getRequestURL(
-  event: H3Event,
-  opts: { xForwardedHost?: boolean; xForwardedProto?: boolean } = {}
-) {
-  const host = getRequestHost(event, opts);
-  const protocol = getRequestProtocol(event);
-  const path = getRequestPath(event);
-  return new URL(path, `${protocol}://${host}`);
-}
